pyproject.toml
[tool.poetry]
name = "eliot"
version = "0.0.1"
description = ""
authors = ["mappland"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
graia-ariadne = {extras = ["standard"], version = "^0.10.0"}
graiax-silkcoder = {extras = ["ffmpeg"], version = "^0.3.5"}
ffmpeg = "^1.4"
aiohttp = "^3.8.3"
httpx = "^0.23.1"
pyyaml = "^6.0"
colorlog = "^6.7.0"
pinyin = "^0.4.0"
pillow = "^9.4.0"
numpy = "^1.24.2"
edgegpt = "^0.0.48"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false
